line wrapping
Hard wrapping is great for code, but not so great for prose.
It makes tweaks and typos get lost in huge diffs when the paragraph is
reformatted.
And many people have softwrapping enabled in the editors for MD files
anyway.

In order to optimize for git usage, markdown files should be hardwrapped
at sentence boundaries.